Correct compiler warnings

+++ src/s57chart.cpp
@@ -4924,8 +4924,7 @@ void s57chart::CreateSENCRecord( OGRFeature *pFeature, FILE * fpOut, int mode, S
             sprintf( line, "  %s %g %g\n", pGeo->getGeometryName(), ref_lat, ref_lon);
             sheader += wxString(line, wxConvUTF8);
         }
-
-        fprintf( fpOut, "HDRLEN=%d\n", sheader.Len());
+        fprintf( fpOut, "HDRLEN=%lu\n", (unsigned long)sheader.Len());
         fwrite(sheader.mb_str(wxConvUTF8), 1, sheader.Len(), fpOut);
 
         if(( pGeo != NULL ) /*&& (mode == 1)*/)
